Major: working version of non-liner solver
This represents the first working version of the parafem non linear
solver for the tutorial case hronTurek. The tutorial test case still
needs to be uploaded however the run scripts need to be updated as
they are copied from the OpenFOAM-Caliux tutorial.

The solver needs cleaning up and implementing in a more sustainable
fashion.